Remove a huge amount of unused code
